requirements.txt
# Web Framework
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
aiofiles==23.1.0
pydantic==1.10.7

# Machine Learning and Data Processing
torch==2.0.0
torchvision==0.15.1
numpy==1.24.2
scikit-learn==1.2.2
Pillow==9.5.0
opencv-python==4.7.0.72

# Utilities
python-dotenv==1.0.0
tenacity==8.2.2
loguru==0.7.0

# Testing
pytest==7.3.1
httpx==0.24.0

# Optional - for development
jupyter==1.0.0
matplotlib==3.7.1
seaborn==0.12.2